Added new column AccountType and used it in CustomerService

+++ src/main/java/com/example/bank/accounts/CurrentAccount.java
@@ -2,7 +2,7 @@ package com.example.bank.accounts;
 
 public class CurrentAccount extends BankAccount {
 
-    public CurrentAccount(int customerId, double balance) {
+    public CurrentAccount(long customerId, double balance) {
         super(customerId, balance);
     }
